# api/oas-doc.yaml
openapi: 3.0.0
info:
  title: School Data Api API
  version: '1.0'
  description: 'This API serves students info, courses info and grades info.'
  license:
    name: ''
  contact:
    name: kihun han
    url: kihun.net
    email: kihun@kihun.net
paths:
...
  '/students/:Id':
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Name:
                    type: string
                  Address:
                    type: string
                  Class:
                    type: string
                  enrolledCourses:
                    items:
                      type: integer
                    oneOf:
                      - type: string
                      - type: array
                        items: {}
                  Id:
                    type: integer
              examples:
                response:
                  value:
                    Name: Micheal Owen
                    Address: Kotkantie 1
                    Class: LIV02
                    enrolledCourses:
                      - 1
                      - 2
                      - 3
                    Id: 1
      summary: Get Student by Id
      description: 'This path will return specific student information with specific id '
      operationId: getStudentById
      parameters:
        - in: path
          name: Id
          description: Unique Student ID
          required: true
          schema:
            type: integer
    put:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Name:
                    type: string
                  Class:
                    type: string
                  enrolledCourses:
                    type: array
                    items:
                      type: integer
                  Address:
                    type: string
              examples:
                response:
                  value:
                    Id: 54350884
                    Name: re
                    Class: ex
                    enrolledCourses:
                      - 79137048
                      - 71583257
                      - 2956846
                      - 21378562
                      - 38107198
                    Address: incididunt commodo aliqua fugiat cupid
      summary: Put Student by Id
      description: 'This path will update specific student information with specific id '
      operationId: putStudentById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Student Id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Id:
                  type: integer
                Name:
                  type: string
                Class:
                  type: string
                enrolledCourses:
                  type: array
                  items:
                    type: integer
                Address:
                  type: string
        x-examples:
          application/json:
            Id: 54350884
            Name: re
            Class: ex
            enrolledCourses:
              - 79137048
              - 71583257
              - 2956846
              - 21378562
              - 38107198
            Address: incididunt commodo aliqua fugiat cupid
    delete:
      responses:
        '200':
          description: ''
          content:
            text/html:
              schema:
                type: string
            text/plain:
              examples:
                response:
                  value: Successfully deleted
      summary: Delete Student by Id
      description: This path will delete specific student data with specific Id
      operationId: deleteStudentById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Student Id
          schema:
            type: integer
...
  '/courses/:Id':
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Name:
                    type: string
                  Description:
                    type: string
                  enrolledStudents:
                    type: array
                    items:
                      type: integer
              examples:
                response:
                  value:
                    Id: 103359
                    Name: adipisicing et
                    Description: cupidatat anim fugiat voluptate
                    enrolledStudents:
                      - 38216
                      - 83472
      summary: Get Course By Id
      description: This path will return specific course data with specific id
      operationId: getCourseById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Course Id
          schema:
            type: integer
    put:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Name:
                    type: string
                  Description:
                    type: string
                  enrolledStudents:
                    type: array
                    items:
                      type: integer
              examples:
                response:
                  value:
                    Id: 52221627
                    Name: sit
                    Description: laborum cillum eu
                    enrolledStudents:
                      - 3573793
                      - 8427
                      - 79892712
                      - 82495333
      summary: Put Course By Id
      description: This path will update specific course
      operationId: putCourseById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique course Id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Id:
                  type: integer
                Name:
                  type: string
                Description:
                  type: string
                enrolledStudents:
                  type: array
                  items:
                    type: integer
        x-examples:
          application/json:
            Id: 97882466
            Name: amet sunt eu dolore
            Description: voluptate dolore deserunt consequat nostrud
            enrolledStudents:
              - 203933
    delete:
      responses:
        '200':
          description: ''
          content:
            text/html:
              schema:
                type: string
              examples:
                response:
                  value: Successfully deleted
      summary: Delete Course By Id
      description: This path will delete the course data with specific id
      operationId: deleteCourseById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Course Id
          schema:
            type: integer
  '/courses/:Id/addStudents':
    post:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Name:
                    type: string
                  Description:
                    type: string
                  enrolledStudents:
                    type: array
                    items:
                      type: integer
              examples:
                response:
                  value:
                    Id: 1665905
                    Name: anim
                    Description: adipisi
                    enrolledStudents:
                      - 89937
                      - 61
                      - 63252371
                      - 391601
                      - 248436
      summary: Enroll Students to Course
      description: This path will enroll the students to course data
      operationId: postStudentsToCourse
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Course Id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Student:
                  type: integer
        x-examples:
          application/json:
            Student: 66851039
  '/courses/:Id/students':
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            text/html:
              examples:
                response:
                  value: 'Enrolled students are student1, student2, student3'
      summary: Get Students By Course Id
      description: This path will return enrolled students list of course
      operationId: getStudentsByCourseId
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique course Id
          schema:
            type: integer
...
  '/grades/:Id':
    get:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Student:
                    type: integer
                  Course:
                    type: integer
                  Grade:
                    type: integer
              examples:
                response:
                  value:
                    Id: -59090642
                    Student: -55241577
                    Course: 20607187
                    Grade: -16604176
      summary: Get Grade By Id
      description: This path will return specific grade data with specific id
      operationId: getGradesById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique grade Id
          schema:
            type: integer
    put:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  Id:
                    type: integer
                  Student:
                    type: integer
                  Grade:
                    type: integer
                  Course:
                    type: integer
              examples:
                response:
                  value:
                    Id: -35288262
                    Student: -46482013
                    Grade: -73643820
                    Course: 2157465
      summary: Put Grade By Id
      description: This path will update specified grade data
      operationId: putGradeById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique grade ID
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Id:
                  type: integer
                Course:
                  type: integer
                Student:
                  type: integer
                Grade:
                  type: integer
        x-examples:
          application/json:
            Id: -31627617
            Course: -95618793
            Student: -77027899
            Grade: 8265773
    delete:
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
            text/html:
              examples:
                response:
                  value: Successfully deleted
      summary: Delete Grade By Id
      description: This path will delete specified grade
      operationId: deleteGradeById
      parameters:
        - in: path
          name: Id
          required: true
          description: Unique Grade ID
          schema:
            type: integer
servers:
  - url: 'http://localhost:3000/'